==> 00_ajustes_sgt_movimentos.ipynb <==
## Problema
'codigo', 'descricao' e 'cod_pai' são nomes genéricos que não auxiliam na codificação, ou qualificam adequadamente a informação a que se referencia. 

## Solução
Utilizar nomes que contextualizem a tabela a que a coluna pertence.

Exemplos:
- cod -> cod_assunto
- descricao -> descricao_classe
- cod_pai -> cod_movimento_pai

import pandas as pd 

data_path = 'dados/01_tabelas_auxiliares/'

# Carrega tabebla base e renomeia os campos
movimento = pd.read_csv(data_path + 'sgt_movimentos.csv', sep=';', usecols=['codigo', 'descricao', 'cod_pai'])
movimento.rename(columns = {'codigo':'cod_movimento', 'descricao':'descricao_movimento', 'cod_pai':'cod_movimento_pai'}, inplace = True)
# Garante que os campos de 'join' possuem mesmo tipo
movimento['cod_movimento'] = movimento['cod_movimento'].astype(float)

# Carrega tabela pai
movimento_pai = movimento.loc[:,('cod_movimento', 'descricao_movimento')]
movimento_pai.rename(columns = {'cod_movimento':'cod_movimento_pai', 'descricao_movimento':'descricao_movimento_pai'}, inplace = True)

# Concatena tabelas e salva
tabMovimento = pd.merge(movimento, movimento_pai, how='left', on='cod_movimento_pai')
tabMovimento.to_csv(data_path + 'tabMovimento.csv', index=False)

print(tabMovimento.shape)
print(tabMovimento.dtypes)
print(tabMovimento)
