==> 00_extracao_dados_full.ipynb <==
## Problema
Para a solução proposta, trabalhar com o formato JSON não é adequado.

## Solução
Extrair os dados para o formato tabular aberto '.CSV'.

import os
import pandas as pd 
import json_assistente as ja

# Gerar arquivos .CSV dos equivalentes .JSON
def gerar_csv(diretorio):

    for file in os.listdir('dados/' + diretorio + '/'):
        print('Extraindo arquivo: ' + file)
        file_name = file.split('.')
        # Efetua extração do JSON
        df = ja.json_file_to_dataframe('dados/' + diretorio + '/' + file)
        df.to_csv('dados/01_tabelas_extraidas/' + diretorio + '/' + file_name[0] + '.csv', index=False)
        print('Finalizada extração: ' + file)
        print()
        
# Gerar os dados em CSV
gerar_csv('00_justica_federal/processos-trf1')
gerar_csv('00_justica_federal/processos-trf2')
gerar_csv('00_justica_federal/processos-trf3')
gerar_csv('00_justica_federal/processos-trf4')
gerar_csv('00_justica_federal/processos-trf5')

print('Extração finalizada!!!')
