==> 02_1_concatenar_tabelas.ipynb <==
## Problema
Trabalhar com loops em arquivos dentro de um diretório é mais trabalhoso, e computacionalmente mais caro.

## Solução
Feita a primeira melhoria nos dados, junta-se todos os dados de um determinado tribunal em um único arquivo.

import pandas as pd 
import numpy as np
import os
import fnmatch

# Une todas as tabelas em uma única
def concatenar_tabelas(diretorio):

    local_dados = 'dados/02_tabelas_transformadas/' + diretorio
    tabFull = None
    
    print('Concatenação inciada: ' + local_dados)
    
    for file in fnmatch.filter(os.listdir(local_dados), 'processos-*.csv'):
        file_name = file.split('.')
        print('Concatenando tabela: ' + local_dados + file)
        if tabFull is None:
           tabFull = pd.read_csv(local_dados + file, low_memory=False) 
        else:
            tabFull = pd.concat([tabFull,  pd.read_csv(local_dados + file, low_memory=False)], sort=False)
        print(tabFull.shape)

    print('Salvando tabela...')
    tabFull.to_csv(local_dados + 'processos_full.csv', index=False)
    print(tabFull.shape)

concatenar_tabelas('00_justica_federal/processos-trf1/')
concatenar_tabelas('00_justica_federal/processos-trf2/')
concatenar_tabelas('00_justica_federal/processos-trf3/')
concatenar_tabelas('00_justica_federal/processos-trf4/')
concatenar_tabelas('00_justica_federal/processos-trf5/')
print("Finalizada concatenação das tabelas!!")
